requirements.txt
google-adk>=1.27.0
fastapi>=0.115.0
uvicorn[standard]
sse-starlette
rembg[cpu]>=2.0.72
Pillow>=10.0
python-dotenv
pydantic>=2.0
